driving_uncertainty/requirements.txt
absl-py==0.9.0
anykeystore==0.2
apex==0.1
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bdlb==0.0.1
bleach==3.1.1
cachetools==4.0.0
captum==0.2.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
cryptacular==1.5.5
cycler==0.10.0
Cython==0.29.17
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
docopt==0.6.2
dominate==2.4.0
entrypoints==0.3
future==0.18.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
grpcio==1.26.0
h5py==2.10.0
hupper==1.9.1
idna==2.8
imageio==2.6.1
imgaug==0.4.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonpickle==0.9.6
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kaggle==1.5.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mock==3.0.5
munch==2.5.0
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
onnx==1.7.0
onnxruntime==1.3.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
ort-gpu-nightly==1.3.1.dev202007102
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.1
PasteDeploy==2.0.1
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
plaster==1.0
plaster-pastedeploy==0.7
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.3
protobuf==3.11.2
ptyprocess==0.6.0
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyparsing==2.4.5
pyramid==1.10.4
pyramid-mailer==0.15.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
sacred==0.7.4
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.13.0
SQLAlchemy==1.3.11
tensorboard==2.1.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
tensorflow-metadata==0.21.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tfp-nightly==0.10.0.dev20200127
torch
torchvision
tornado==6.0.3
tqdm==4.47.0
traitlets==4.3.3
transaction==2.4.0
translationstring==1.3
typing-extensions==3.7.4.2
urllib3==1.24.3
velruse==1.1.1
venusian==3.0.0
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==0.16.1
wrapt==1.11.2
WTForms==2.2.1
wtforms-recaptcha==0.3.2
zipp==3.0.0
zope.deprecation==4.4.0
zope.interface==4.7.1
zope.sqlalchemy==1.2
